refactor: replace JSX with String.raw template in getWebviewHtml
